split share sidebar in smaller single components
