+++ src/js/components/Shares/PublicShareItem.vue
